requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow-cpu
matplotlib
joblib
ta
yfinance
